requirements.txt
# PiRadio 의존성
python-mpd2>=3.1.0
yt-dlp>=2024.1.0
ytmusicapi>=1.5.0
flask>=3.0.0
flask-socketio>=5.3.0
pyyaml>=6.0
Pillow>=10.0.0
requests>=2.31.0
schedule>=1.2.0
gevent>=24.2.1

# 라즈베리파이 전용 (Pi에서만 설치)
# st7789>=0.0.4
# spidev>=3.6
# gpiozero>=2.0
# lgpio>=0.2
# numpy>=1.24.0
